Add option to run performance tasks with security disabled
It's noted that kernel fixes for security/meltdown
has affected performance at least I/O. As we run deployments
in different environments it's good to add an option
which can adjust the security settings to test performance.

This patch adds a new variable "kernel_sec" which
defaults to "true". This can be set to "false" in
order to run performance tasks with disable security.

Also add molecule scenario to test security_disable tasks.

Change-Id: If6f0bc02db6932aa8f367e38cc82df85063eaf87
Related-Bug: #1757556