Remove any existing nameserver entries before adding IPA server
If the undercloud nodes is created with entries already
in the /etc/resolv.conf file, the entries added to point to
the IPA server will be added after the existing entries.
This will cause the iIPA enrollment of the client to fail.

This review removes the existing entries before adding the
search domain to avoid the enrollment errors.

Change-Id: I67ebe6c429e89f6f303bf23ca5959e00d6a46df8