Centralize tempest execution
Today we have two roles that run tempest, the first one that is
deprecated, validate-tempest, and the new one os_tempest.
The problem right now is that, in order to deprecate validate-tempest,
we need to ensure a smooth transition, and centralize everything in one
single playbook, is the best way to do it.
Furthermore, move old jobs running on centos 7 require some workarounds
to create the clouds.yaml file, that today would be harder.
This patch consolidate both os_tempest and validate-tempest in one
single playbook, and so, in the future, it's easier to just remove the
calls from validate-tempest in one single place rather than keep
searching for all the places we use it.

Change-Id: I9991f6c0ea51a3bf537bdcafd4220a5a025cc026