Remove install_skiplist and set python interpreter
The install_skiplist role here is not necessary since tripleo-quickstart
already install the openstack-tempest-skiplist in the virtualenv.
This also ensure that the ansible_python_interpreter is set to the
virtualenv one, because ansible.cfg have the interpreter_python set to
auto, and it was executing this step with the system python, making the
import of tempest_skip fail.

Change-Id: Ia4c84ac4c93466afde68abea19e18ca738e99910