Add delegate_to in skiplist tasks
The skiplist requires python module from openstack-tempest-skiplist
available in the python path. This works when quickstart is being
executed from the undercloud, but in some systems where quickstart is
executed in a different host, it fails. Use delegate_to ensure it
will be executed from the host quickstart is being executed.

Partial-Bug: #1888244

Change-Id: I433e39c212e4ab887b6ed0776715e344bfa404d7