Generalize vars in template file in build-test-packages role
With this patch we are generalizing vars in template file to take care
of differenet osp_release naming conventions we currently have in 
downstream:-

For 17:- osp_release: rhos-17
For 16.2:- osp_release: rhos-16.2

We have plans to switch 17 osp_release to numbered in downstream from
osp_release: rhos-17 --> osp_release: rhos-17.0 in near future so that
these kind of changes are not required. 
But as osp_release for 17 is in multiple places, For now we are going 
ahead with this change to make both versions working.

Change-Id: I101d1d938862785ec7afc5d5a7c97ad2f1e78c04