Fetch non tripleo containers in provider job and consume it
When provider_job is set to true, it will pull the container
images from quay.io and will push it to local
registry using non_tripleo_containers.yml.

When consumer_job is set to true in child job, then
custom_ceph_parameters will be setted to true and ceph containers
related namespace will be changed to local registry
to pull the ceph containers so that container built and pushed
to local registry in provider job, will be consumed here.

It also adds standalone job which depends on content provider
job and updates the ceph tag.

Setting +standalone_container_ceph_updates to false as with in
tasks it is setted default to false.

Moved local registry host and port to extra-commons role
to easily share between roles.

It also fix the non_tripleo_containers playbook to exclude kolla
containers for train release.

Note to self:
tripleo-standalone-scenarios-pipeline will be added later
once standalone-upgrade job is ported to centent provider.

Change-Id: Ie3f7bafe3b6ee8a7879ebb3d1d5b47b20181b452
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Co-authored-by: Alex Schultz <aschultz@redhat.com>