Add the variable list_skipped_job
The list_skipped role have a variable called list_skipped_job.
This variable makes the role filter the list of tests to be
skipped per job. This makes required in order to start use
the tempest-skiplist to skip tests for specific jobs (for
example component-pipeline jobs).
Without passing this variable, we are not able to filter
tempest tests per job.

Depends-On: https://review.opendev.org/#/c/751828/
Change-Id: I640b4efc9474614b44cbcf56d52736760dffc230