Set container namespace to local registry in consumer job
Current in container-prep-parameters file, namespace is pointing
to docker.io in consumer job but it needs to be pointing
to local registry to pull containers from provider job when
consumer_job is set to true.

Added local_registry_ip_address and port to the default vars to
avoid vars not defined issue.

It fixes the same.

Change-Id: Iaa2b3b0074732287b618fe268a7e8876f14c6348
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>