Standalone configure neutron bridge correctly
This change updates the default bridge mapping from datacentre:br-ex to
datacentre:br-ctlplane. We're doing this because in the standalone in
CI, we configure a br-ex before running the standalone (via
undercloud-setup) and want to attach our br-ctlplane to it. We then want
to ensure that we use br-ctlplane for the neutron access to the external
network to prevent weird routing issues when we have two bridges on the
same subnet.

Depends-On: https://review.opendev.org/#/c/757605/
Change-Id: I0e5aa3f58746dc0b92bd35ade7792f323b5647f7
Related-Bug: #1895822