Map build host repos to distro.repos.d instead of yum.repos.d
Currently, when building container images, we are overriding the ubi
image repos with the repos configured in the host. Effectively, this is
removing ubi repo config from the image build environment which most of
the times is not a problem as centos is on sync with ubi, but in some
cases as after RHEL 8.3 release, ubi is ahead of CentOS and need
packages from the ubi repo itself.

This patch is bind mounting the repos config from build host in
distro.repos.d instead of yum.repos.d. In that way, both the repos
configured by default in the image (ubi repos) and the ones in the host
are available.

Depends-On: https://review.opendev.org/761702
Change-Id: I08d44d9811bd2666f6732c0541bbdd4a0345d265
Closes-Bug: #1902846