Enable the component-ci-testing repo if it is installed
This change enables the component-ci-testing repo if it is
present. This must happen after the repo-setup role is called
otherwise that will overwrite any changes made.

For component jobs we are injecting the component-ci-testing
repo with add_repos - for example at [1]. For upgrade jobs we
must install this as disabled, until the upgrade part of the
job otherwise we are confusing repo versions as explained
at [1]. That change depends on this one.

[1] https://review.rdoproject.org/r/#/c/31200/2/zuul.d/component-jobs.yaml@1225

Change-Id: I2b82c2be49c496010e09bc9df48ca826b6fab66d