Set properly the rc file to create the clouds.yaml
The previous releases (queens, rocky and stein), don't have the
clouds.yaml file, and in order to run os_tempest this is required. In
these releases we write the clouds.yaml manually to pass to the
os_tempest, however, the logic was a little wrong because it was done
only when tempest is running in the overcloud. In the case of undercloud
it is failing. This patch fix it setting the rc file properly (stackrc
or overcloudrc).
Also, in order to avoid errors when tempest run on undercloud, set the
variable tempest_service_available_neutron to false.

Change-Id: I748c3eb7a4dc299e5368a1330ad22fa48de912a3