Trigger validations only on train and above.
In patch[1] we have updated the playbook to use new validation role
'tripleo_validations' stored in tripleo-validations repo.

Above patch breaks stein and older branches as that role is not
available prior to train branch.

With this patch we skip the include of new tripleo_validation role for
stein and prior branches.

We also moved to include_role style way of including role so that if
role is not available it do not try to load it.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/755382
Closes-Bug: #1914993

Change-Id: Ida95fd46718e5d29d9951453653afb390b0d7bbd