Enable nested virt for scenario 10 octavia kvm job
For sc10 kvm jobs, we need to enable nested virt on the host.
It adds the execution point for the same.

It cannot be enabled by default for sc10 kvm jobs as
not all upstream cloud support nested virtualization.

Earlier sc10 kvm jobs were not added to the pipeline.
We are working on enabling it here:
https://review.rdoproject.org/r/q/topic:%22fix_octavia_kvm%22+(status:open%20OR%20status:merged)

and here is the test project change:
https://review.rdoproject.org/r/c/testproject/+/32573

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ida4349a1f0fdee5dce0ec2bbc96d0567c6d075d0