Collect and output outputs/passwords files with old and new locations
In Idded7faba1ff6c811b94503c559029aeeaca6a06, we are changing the
default path for tripleo-undercloud-outputs.yaml and
tripleo-undercloud-passwords.yaml. These tasks in CI that expect the
files to exist at the old location are updated to use the first_found
lookup to check both the old and new paths.

Likewise, when the output files are written on the minion, they are
written to both the old and new locations until the tripleoclient patch
merges.

Once the tripleoclient patch merges, we can update these tasks to only
look at the new path.

Partial-Bug: #1921975
Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: If28d5b225439162b7983165fe3bea605971cfb74