Remove tripleo packages from repoclosure exception
In order to find out dependencies resolution conflict with in
a repo, repoclosure is added.
Initially we added tripleo packages to exception list as it
was failing earlier as we were not checking it properly.

delorean-tripleo-component comes from delrean.repo and release
file also layout delorean-current repo with including packages
which overrides the packages config leading to packages conflicts.

As both the repos are same, initially we disable all repos and
enable all delorean repos but we need to disable only delorean
current repo to avoid overriding due to include packages in the
end.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I55ede9aa7741b0fa3604c97bd42c43a22f1d1fe3