Set flavor_min_mem as a tempestconf overrides
For older release like queens and stein, tempestconf-2.4.0
is used and in all the branches we are using cirros-0.5.2
image which requires DEFAULT_FLAVOR_RAM to 128 not 64( which
is defined in older release of tempestconf).

We cannot bump the tempestconf version to latest as older
branches depends on py2 and tempestconf uses py3.

In order to fix the issue, we can pass the
flavor_min_mem value via tempestconf profile to 128 to
fix the issue.

https://review.rdoproject.org/r/c/testproject/+/33649
getting tested here.

Closes-Bug: #1927739

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I21824bc14f4c1d55235d7b80870bac7f7edff5d4