Update public subnet allocation pool
Previously allocation pool was starting with tempest_cidr.100
and ended with tempest_cidr.150 IP addresses.
In some jobs it overlapped with the IPs used for
RedisVirtualFixedIPs (192.168.24.101) and
OVNDBsVirtualFixedIPs (192.168.24.102)

In such case, when e.g. 192.168.24.102 was used in tempest test as FIP,
test was failing due to paramiko.ssh_exception.BadHostKeyException
raised as on the ~/.ssh/known_hosts in the undercloud there was already
key for host with IP 192.168.24.102.

Those RedisVirtualFixedIPs and OVNDBsVirtualFixedIPs were added in
commit [1].
This patch changes allocation pool used in tempest to be
tempest_cidr.150 - tempest_cidr.200 to avoid that overlap.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/777106

Closes-Bug: #1928933
Change-Id: Ia515a1523a38a2d634d9520a9d9453218b9c38e1