Add release_numbering & release_minor support
With this patch we are adding support for release_numbering
& release_minor for downstream.

We will pass these parameters through downstream release config
files, should be transparent for upstream.

• `release_numbering` defines the algorithm used by DLRN to assign
   release numbers to packages.

The release number is created from the current date and the source
repository git hash, and can use different algorithms:

– 0.date.hash if the old method is used: 0.<date>.<hash>
– 0.1.date.hash if the new method is used: 0.1.<date>.<hash>.
This new method provides better compatibility with the Fedora
packaging guidelines.
– minor.date.hash allows you to specify the minor version to be
used, which can be different from 0.
If this release numbering schema is used, the value of minor
will be determined by release_minor.

• `release_minor` only takes place when release_numbering
is set to minor.date.hash. For example, if this value is set to 3,
the release number for all packages will be computed as 3.date.hash.

Change-Id: Ie20ccbcbeed0769202a8408e7bae0465daaa1803