Add list_skipped_installer option to tempest playbook
After [1] get merged, it is required to set the list_skip_installer
variable properly to osp and tripleo. Right now, since the default is
tripleo, everything will be working fine. But if you want to distinguish
between osp and tripleo, need to set a ternary if when osp_release is
defined or not, which is exactly what this patch does.

1 - https://review.opendev.org/c/openstack/openstack-tempest-skiplist/+/789625

Change-Id: I8b181d3316ff595f44ff2fae42a380ca68eb6aa1