Utilize t-o-a role for import/introspect in train
We are already using tripleo operator role for import /introspect for
ussuri+ since patches[1] and [2].

With this patch we are starting to use operator role in train branch
as well.

Also fixes the legacy task for introspection, Issue introduced in [3].
`changed_when: true` got over indented under the shell module.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/726847
[2] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/800215
[3] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/800215

Closes-Bug: #1935948

Change-Id: I153c56461d8def736d056e01e17283322a0d621e