Update list_allowed variables
Before, the list_allowed_job and list_allowed_group were mutually
exclusive, however, due the nature of featuresets, in order to remove
the tempest_test_whitelist from featureset files, we need to replace it
for tempest_allowed_group and allow it to pass both to list_allowed
role, that right now already update the logic to in case both group and
job are passed, job have precedence over group. So, if a featureset have
a group set, and there is no specific job with specific tests, it will
use the list tempest tests, otherwise, it will use the job name.

Change-Id: I90530cca3b7195241646972d98a180a025dbd7e2