Replace undercloud-setup IPA part with specific tasks
undercloud-setup rols is used by ipa-multinode role while
setting up ipaserver undercloud.

Not all the tasks of undercloud-setup is used there and
recently we moved the build-test-packages as a part of
undercloud-setup and added condition[1] specific to ipa
for skipping build-test-packages.

Which leads to skip fs039 (ovb based IPA deployment)
Run DLRN gate role tasks which calls build-test-packages
role.

Here is the example from the job:
https://review.rdoproject.org/zuul/builds?job_name=periodic-tripleo-ci-centos-8-ovb-3ctlr_1comp_1supp-featureset039-master
```
2021-09-02 01:48:42.859718 | primary | TASK [Run DLRN gate role] ******************************************************
2021-09-02 01:48:42.859779 | primary | Thursday 02 September 2021  01:48:42 +0000 (0:00:00.122)       0:01:08.801 ****
2021-09-02 01:48:42.955410 | primary | skipping: [undercloud]
2021-09-02 01:48:42.970685 | primary |
2021-09-02 01:48:42.970728 | primary | TASK [Install gated repo] ******************************************************
2021-09-02 01:48:42.971064 | primary | Thursday 02 September 2021  01:48:42 +0000 (0:00:00.111)       0:01:08.912 ****
2021-09-02 01:48:43.086896 | primary | skipping: [undercloud]
```

In order to avoid that, we are reusing specific tasks of
undercloud-setup in ipa-multinode role to fix the issue.

[1].https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/804971

Testproject: https://review.rdoproject.org/r/c/testproject/+/35104

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ifba29f9045d4e7e0d6b5ef1446e4ce4e28b44cad