Correct overcloud delete fail condition
Currently, We are looking for string "overcloud" in
`tripleo_overcloud_status_output` to confirm if overcloud is deleted
or not.

Even when overcloud is present, we get "overcloud" in output and
hence we can miss some conditions where overcloud is not
deleted successfully.

~~~
$ openstack overcloud status
+------------+-------------------+
| Stack Name | Deployment Status |
+------------+-------------------+
| overcloud  |   DEPLOY_SUCCESS  |
+------------+-------------------+
~~~

If overcloud is not available, We get:-

~~~
$ openstack overcloud status
No deployment was found for overcloud
~~~

"No deployment was found" string will be better filter to catch
if overcloud is deleted or not.

Change-Id: I41283afbcd81226b9322696d6d5265077684e2be