Override inventory var value if the job defines it's own
The default value of the variable 'inventory' of the validations.yml
playbook remains the same. However, if the zuul job defines alternative
inventory for the VF to use, it will be used instead.

Without exposing the variable to the job, the inventory path is hard set
and can not be adjusted to match exact location of the inventory.
This poses problem for validation execution.
As VF relies on inventory information to match validations to hosts.

Also removes conditional for older releases.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ia15063cabc65e2ae685e2f9dcaec9d25b05b4db8