Revert "Revert "Use tripleo-mount-image for modify-image""
This reverts commit a7a05c4371bf894a58b63622a7995e6e4ee31098.

This was failing on train, ussuri, victoria because the fallback path
assumed the overcloud-full partition image is in raw format, but is
qcow2 on these stable branches.

This revert now includes the same image format detection that was
introduced to tripleo-mount-image in change
I628abc4cb8c46c3b2e5bfd3cd16bb6d5cc95fabe.

Change-Id: I0354a7f8804949916cfccfc22c59f26b54c4c134