Set default whole_disk_images based off release
There is a need to stop setting this in the releases file so that some
jobs in the release can have a different value (See
I0b555676ce565d142a08e9d38d9e3c8ee7cb1a92). This change sets the
default based off the release.

It needs to be determined if the release config in
tripleo-quickstart[1] picks up on this default setting so that
whole_disk_images:true can be reliably set in the required jobs.

[1] https://opendev.org/openstack/tripleo-quickstart/src/branch/master/config/release/tripleo-ci/CentOS-8/master.yml#L33

Change-Id: I1e305d31acba2f2b49fca675a0c6cb4eec5c5946