Condition to create n/w: custom overcloud name
Currently we are creating networks only when tempest_cloud_name
is i.e standlone or overcloud.

We are creating a new job with custom overcloud name which is currently
failing because we are not creating public network because
tempest_service_available_neutron is false and condition [0]

Updating condition to incorporate custom overcloud name.

[0] https://opendev.org/openstack/openstack-ansible-os_tempest/src/branch/master/tasks/tempest_resources.yml#L209
[1] https://opendev.org/openstack/openstack-ansible-os_tempest/src/branch/master/tasks/tempest_resources.yml#L188

Change-Id: Ie7c6db14318649f6dfb9f3d765dbe462697cc82a