Inject gating gating repo during image build
Currently we inject gating.repo on the built image.
In fs02, we first built the image and then inject
the gating repo on the image to pick up latest changes.

On FS02, the gating repo is not getting applied on the
built images.

In order to fix that, we are including the gating.repo
during the image build process and the latest packages
are getting pulled from gating repo.

It also removes install built repo operation as build
images role already takes care of that.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I58ffeafddac3e14e265fc19f294f4508b8f7f5e9