Add mirror_info.sh creation for supplemental node repo setup
As discussed in related-bug the supplemental node used in the IPA
jobs is not using the mirror_info.sh so not using mirrors in the
repos created by repo-setup. This adds a slurp to grab the mirrors
config from the undercloud before repo-setup on supplemental.
Test and results at [1].

[1] https://review.rdoproject.org/r/c/testproject/+/40286/1#message-f825e6369163ad1f5aa98eb7c6c67b2798d5788b
Related-Bug: 1964457
Change-Id: I7b3c6f0a3b9fa7114670656df7b39cef8047fd85