Added comment to revert in a future
The issue which was solved doesn't have a correct solution yet.
It might be tricky and it's not clear when solution will be
available.
IPA server does not work correctly under latest openssl.
Due to that, it's recommended to keep pinned version of
openssl to known working one.

Related-Bug: https://bugzilla.redhat.com/show_bug.cgi?id=2057471
Closes-Bug: 1963545
Change-Id: Ie037697098d53b8c1c49fa5e6679e244e92ac768