Override image.http_image on tempestconf
The web-download test expect an url in order to work, otherwise it fails
with an error saying unable to parse the http_image.
Tempestconf set the http_image to the same as image_path if http_image
is not overwriten, so, the http_image was pointing to
/opt/cache/files/cirros. This patch set the http_image to the proper
cirros url. I check the tempest code, and only the web_download test
seems to be using the http image, so it should be safe to override the
http_image here.

Change-Id: I085177212a6e9b461009c453f7907867b118f859
Closes-Bug: #1965540