Disabling atop installation by default
The atop package is meant to be installed during the undercloud setup
for diagnostic purposes. Unfortunately it is not available in Centos8 Stream
or Centos 9[0] which are used as OS by many of our CI jobs.

The package is also not present in package manifests for RHEL 8[1] and RHEL7[2].

Taken together, this makes the task installing the package consistently fail.

While the failure is ignored, the task does query the package manager
and the output pollutes the logs.
Setting the the atop installation to not be run by default is the first
step to remedy the situation.

Closes-Bug: #1968014

[0] https://pkgs.org/search/?q=atop
[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/package_manifest/index
[2] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/package_manifest/index

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I9688c5cc0e3db468a658ee903bb581c3c5d6640f