Convert a hieradata override to plain t-h-t parameter
Since we're moving the masquerade management out of puppet[1], we
cannot rely on hieradata anymore.

This patch creates a custom_env_file taking care of the override of the
actual heat MasqueradeNetworks parameter.

It also get rid of an ugly pattern, and uses an actual list for the
custom_env_files (while still supporting the space separated string).

[1] https://review.opendev.org/q/topic:masquerade-ansible

Depends-On: https://review.opendev.org/c/openstack/tripleo-ansible/+/836871
Change-Id: I977d245eb4863b37c63e21c416debe8df34b813b