Add role to discover Centos latest images
With this patch, we are adding a new role discover-latest-image to
discover the latest Centos Stream image, We are limiting the scope
for C9 for now, We can extend the role for Centos 8 image discovery
in future if there is a need.

Also, we are updating baremetal-prep-virthost.yml playbook to use
this new role for c9 case and in combination with patch[1] - we will
get rid of undercloud baseos hardcoding and use the latest discovered
image.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart/+/838990/

Change-Id: I3a97edc63e96725a1690335f557508d56ed10258