Propagating branch_override variable into DLRN build role
The role was previously set to handle only branch names directly received from
Zuul or preset default. This could lead to role checking out incorrect distgit
repos in cases when the name of the branch in upstream repo differs significantly
from that in the distgit repo.

This change will allow per job override of the behavior, using the 'branch_override' var.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I3c6a8612de60b201b9f4ac1070f52d4debe37eb5