Fix tempest_allow logic
Since we weren't using the tempest-allow role before, the logic was set
to be executed when tempest_test_whitelist wasn't defined, which is
fine, but to run skipped jobs, the allowed list was being overwriten.
This patch add the logic to only run the tempest-allow role when
tempest_run_skipped is set to false (default).

Change-Id: Ie4a5b6a764d19ce49f703804daaa50425be1bb6f