Disable direct override using ironic_config
... because it does not allow us to implement these parameters in
puppet-ironic. We should NOT use this interface in CI but should use
only native puppet parameters.

This replaces usage of arbitrary interface by the native puppet
parameters. These parameters are not yet implemented because
the existing override is causing duplicate definition, but will be
added once this change is merged.

Change-Id: I1754111c943afc004ef3a3226f662d415f910999