Remove podman_version fact and conditional from undercloud-deploy
This was added in [1] but no longer needed. The fact is not consumed
anywhere else [2], the branches mentioned are EOL and podman is now
on v3.0 in train e.g. [3].

The conditional is causing problems for the mixed os job in [4]
because we include_tasks and not the whole undercloud-deploy role to
regenerate the containers prepare file.

[1] https://opendev.org/openstack/tripleo-quickstart-extras/commit/9d31d8996cb8a92bf9b30e90096c4222c1ca0779
[2] https://codesearch.opendev.org/?q=podman_version
[3] https://logserver.rdoproject.org/openstack-periodic-integration-stable4/opendev.org/openstack/tripleo-ci/master/periodic-tripleo-ci-centos-8-containers-multinode-train/5dc578c/logs/undercloud/var/log/extra/package-list-installed.txt.gz
[4] https://logserver.rdoproject.org/34/44234/1/check/periodic-tripleo-ci-centos-8-9-multinode-mixed-os/2892353/job-output.txt

Change-Id: Iab4776bb71e37bacae77e678484b2a1f308573db