Add custom CI rules to the Undercloud Firewall
We noticed some packet drops on undercloud node after switching to
nftables [1] and [2]

In order to properly get VXLAN support, We have to add a couple of
rules to the Undercloud firewall. This is especially true for a lab
environment, or on the upstream CI infrastructure.

Also, Port 19885 is for console log streaming.

Patch[3] add this requirement in the documentation as well.

Testproject[4] works well, We can check the rendered
undercloud-parameter-defaults.yaml[5] its is correctly poppulated
with right rules.

[1] https://logserver.rdoproject.org/54/31954/58/check/periodic-tripleo-ci-centos-9-ovb-3ctlr_1comp_1supp-featureset039-master/490510a/logs/undercloud/var/log/extra/dropped-packets.txt.gz
[2] https://logserver.rdoproject.org/54/31954/49/check/periodic-tripleo-ci-centos-9-containers-multinode-master/66bb894/logs/undercloud/var/log/extra/dropped-packets.txt.gz
[3] https://review.opendev.org/c/openstack/tripleo-docs/+/851849
[4] https://review.rdoproject.org/r/c/testproject/+/31954/61#message-6f6b2fe65899192fd18eb110665d2361bd38682d
[5] https://logserver.rdoproject.org/54/31954/61/check/periodic-tripleo-ci-centos-9-ovb-3ctlr_1comp-featureset001-master/27d2796/logs/undercloud/home/zuul/undercloud-parameter-defaults.yaml.txt.gz

Change-Id: Ic19ed9b8e95dbe2d3ad7efce7221a28dc8526be4