Update CEPH_NAMESPACE and rely on the "release" variable
Now that master is on Quincy but Wallaby is still on Pacific, we need to
make sure the container namespace is the right one. This patch removes
the hardcoded master container namespace and fetch it according to the
openstack branch.
Note that centos9 can be hardcoded because we test both releases on c9.

Change-Id: I7db138fd012e264e87009c524eee02448e793ddf