Create a yaml with custom CI rules for Firewall
We need some tweaks in the firewall rules in order to be able to switch
to nftables.

After analyzing the dropped packets[1], we must open a couple of ports
related to VXLAN[2], as well as ensuring ssh has the same opening as the
current "openstack-INPUT" chain created in iptables via the CI itself.

This patch generates the needed configurations for both undercloud and
overcloud, since both will basically need the same things.

[1] https://logserver.rdoproject.org/54/31954/69/check/periodic-tripleo-ci-centos-9-scenario007-multinode-oooq-container-master/863a967/logs/subnode-1/var/log/extra/dropped-packets.txt.gz
[2] https://opendev.org/openstack/tripleo-docs/src/branch/master/deploy-guide/source/features/security_hardening.rst#vxlan-and-nftables

Change-Id: Ic3e0f63e98d97323c313a87045d0fe01bec5ded8