Do not use --debug for image build
Since Ansible was bumped to 2.14, we've observed the container image
build process gets stuck in the middle of ansible tasks to generate
Docker/Buildah files, because of a bug[1] with ansible-runner.

This removes --debug option from the build command to avoid -vvv option
in the ansible command, to workaround the above bug.

[1] https://github.com/ansible/ansible-runner/issues/1164

Related-Bug: #1996612
Change-Id: I53c688077c65da03d8c3cf104862e02cefc2c615