Add support for role specific container prep to undercloud-deploy
This adds support for the per-role container image prepare params
into the undercloud-deploy role. See [1] for other related patches.

The undercloud-mixed-os play is changed to resolve the rhel/centos8
ci-testing hash for periodic jobs before undercloud-deploy for the
multi container prep.

Vars are being set with [2] but the single deploy workflow will
not be enabled until [3] merges.

[1] https://review.opendev.org/q/topic:role-image-prepare-stable/wallaby
[2] https://review.opendev.org/c/openstack/tripleo-quickstart/+/868208
[3] https://review.opendev.org/c/openstack/tripleo-ci/+/868207

Change-Id: Ieaad3d17bb9c63f7042f0454d81dffada98c5947