Run container-prep role before installing standalone
In order to set correct container tag from dlrn md5 sum
in container-image-prepare file, we need to run
container-prep role before deploying standalone.

Otherwise tripleo-ci-testing tag[1] will be used in periodic
line and this tag does not exists on quay.rdoproject.org.
It will fail the standalone deployment with image not found.

This patch fixes the same and now tag is properly set in
container-image-prepare[2]. The same tag will be set
in deploy_compute.sh script.

[1]. https://logserver.rdoproject.org/18/46118/3/check/periodic-tripleo-ci-centos-9-standalone-external-compute-target-host-master/1355bd7/logs/undercloud/home/zuul/containers-prepare-parameters.yaml.txt.gz

[2]. https://logserver.rdoproject.org/18/46118/4/check/periodic-tripleo-ci-centos-9-standalone-external-compute-target-host-master/a572928/logs/undercloud/home/zuul/containers-prepare-parameters.yaml.txt.gz

Signed-off-by: Chandan Kumar <chkumar@redhat.com>
Change-Id: I6cdb21b46385017b2c56771e2cf45f340013de02