Pass centos.repo from host during image build.
Image build in component jobs is pulling content from
centos.repo instead of quickstart repos, see [0]

This can cause mismatch of rpm when mirrors.centos.org and
local mirrors are not in sync.

The base image which we use to build overcloud images already have
centos.repo and when proxy mirros are not updated this can cause
an issue.

This is a workaround patch to pass centos.repo(which are disabled
on host) so that same repos in the image will be overridden.

One change in behavior is at the end of image build, dib will delete
the centos.repos in the overcloud image as dib cleans up what it adds
ignoring what was already present.

oooci-build-images already have same fix.

[0] https://logserver.rdoproject.org/openstack-component-baremetal/opendev.org/openstack/tripleo-ci/master/periodic-tripleo-ci-centos-9-ovb-3ctlr_1comp-featureset001-baremetal-wallaby/1cd7430/logs/undercloud/home/zuul/build.log.txt.gz
~~~
2022-12-27 00:54:07.471 |  kernel                         x86_64  5.14.0-214.el9            baseos           2.8 M
~~~

[1] https://github.com/openstack/tripleo-ci/commit/b2ce4b4c101d68eaa71d1499609423e10f50f2d5

Related-Bug: #2000226
Change-Id: Iecf36eff8ef27fd2734f77b09180d0b4d8654c52