Remove openstack-tox-molecule job
This job is failing due to some updates in setuptools and there is
no reason to run it, since it's testing only two roles that have not
being update in the last 3 years. Basically, setuptools complains to
install PyYAML because it has on version 5.4.1 the option license_file
that was deprecated. Updating PyYAML also requires to update all
molecule packages, that introduce different failures, in a snowball
effect.

Change-Id: Id5c21dcea9d183c7fac347323893415f4620cee5